Remove jwt_helper.go
